requirement.txt
astroid==2.4.1
click==7.1.2
colorama==0.4.3
Flask==1.1.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.20
protobuf==3.11.3
pylint==2.5.2
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
